Split oversized spec files into module-tests directories
Reorganize five oversized spec files into {module}-tests/ directories,
grouping describe/it blocks by domain concern and extracting shared
fixtures into co-located helpers.ts files. No behavior change — tests
were moved verbatim and the total test count is unchanged.
