--- src/lib/game/modes/werewolf/utils/win-condition.spec.ts
@@ -1,820 +1,0 @@
-import { describe, expect, it } from "vitest";
-
-import type { Game } from "@/lib/types";
-import { GameMode, GameStatus, ShowRolesInPlay } from "@/lib/types";
-
-import { WerewolfRole } from "../roles";
-import { DEFAULT_WEREWOLF_TIMER_CONFIG } from "../timer-config";
-import type { WerewolfTurnState } from "../types";
-import { WerewolfPhase } from "../types";
-import { checkWinCondition, WerewolfWinner } from "./win-condition";
-
-function makeGame(
-  roleAssignments: { playerId: string; roleDefinitionId: string }[],
-  turnState?: WerewolfTurnState,
-): Game {
-  return {
-    id: "g1",
-    lobbyId: "l1",
-    gameMode: GameMode.Werewolf,
-    status: turnState
-      ? { type: GameStatus.Playing, turnState }
-      : { type: GameStatus.Playing },
-    players: roleAssignments.map((a, i) => ({
-      id: a.playerId,
-      name: a.playerId,
-      sessionId: `s${String(i)}`,
-      visiblePlayers: [],
-    })),
-    roleAssignments,
-    configuredRoleBuckets: [],
-    showRolesInPlay: ShowRolesInPlay.None,
-    ownerPlayerId: "owner",
-    modeConfig: {
-      gameMode: GameMode.Werewolf,
-      nominationsEnabled: false,
-      trialsPerDay: 1,
-      revealProtections: true,
-      hiddenRoleCount: 0,
-      showRolesOnDeath: true,
-      autoRevealNightOutcome: true,
-    },
-    timerConfig: DEFAULT_WEREWOLF_TIMER_CONFIG,
-  };
-}
-
-function makeDayTurnState(
-  overrides: Partial<WerewolfTurnState> = {},
-): WerewolfTurnState {
-  return {
-    turn: 2,
-    phase: { type: WerewolfPhase.Daytime, startedAt: 1000, nightActions: {} },
-    deadPlayerIds: [],
-    ...overrides,
-  };
-}
-
-describe("checkWinCondition", () => {
-  describe("Village wins", () => {
-    it("Village wins when all Bad players are dead and no Neutral remain", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-
-    it("Village wins when all Bad are dead and a non-killer Neutral (Tanner) is still alive", () => {
-      // Non-killer Neutrals (Tanner, Executioner, Dracula, Zombie) are not a threat to
-      // Village win — only killer Neutrals (Chupacabra, Arsonist) delay the Village win.
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Tanner },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-
-    it("Village wins when all Bad are dead and an Executioner is still alive alongside Good", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Executioner },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-  });
-
-  describe("Werewolves win", () => {
-    it("Werewolves win when Bad count equals non-Bad count", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Werewolves);
-    });
-
-    it("Werewolves win when Bad count exceeds non-Bad count", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Werewolves);
-    });
-
-    it("game continues when Bad count is less than non-Bad count", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-  });
-
-  describe("Chupacabra", () => {
-    it("Chupacabra wins when all Bad are dead and ≤1 Good remains", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Chupacabra);
-    });
-
-    it("Chupacabra wins when all Bad are dead and 0 Good remain", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1", "p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.Chupacabra);
-    });
-
-    it("game continues when Chupacabra is alive and >1 Good remain", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result).toBeUndefined();
-    });
-
-    it("game continues when Bad and Chupacabra are the only survivors (conflicting win conditions)", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // p3 (Good) is dead — only wolf vs chupacabra remain
-      const result = checkWinCondition(game, ["p3"]);
-      expect(result).toBeUndefined();
-    });
-
-    it("game continues with multiple Bad vs single Chupacabra and no Good", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // p4 (Good) is dead — 2 wolves vs chupacabra
-      const result = checkWinCondition(game, ["p4"]);
-      expect(result).toBeUndefined();
-    });
-
-    it("Village does not win while Chupacabra is alive (with Good players remaining)", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).not.toBe(WerewolfWinner.Village);
-    });
-  });
-
-  describe("Lone Wolf", () => {
-    it("Lone Wolf wins when only Lone Wolves remain vs non-Bad (no regular wolves)", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.LoneWolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.LoneWolf);
-    });
-
-    it("Lone Wolf wins when multiple Lone Wolves outnumber non-Bad", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.LoneWolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.LoneWolf },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.LoneWolf);
-    });
-
-    it("game continues when Lone Wolf alone is outnumbered by non-Bad", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.LoneWolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("regular Werewolves win when both regular wolf and Lone Wolf are alive and meet threshold", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.LoneWolf },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Werewolves);
-    });
-
-    it("Lone Wolf alone does not trigger regular Werewolves win", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.LoneWolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).not.toBe(WerewolfWinner.Werewolves);
-    });
-  });
-
-  describe("Spoiler", () => {
-    it("alive Spoiler overrides Village win", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-      ]);
-      // Werewolf dead → Village would normally win, but Spoiler is alive
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Spoiler);
-    });
-
-    it("alive Spoiler overrides Werewolf win", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-      ]);
-      // Werewolves >= non-Bad → Werewolves would normally win, but Spoiler is alive
-      const result = checkWinCondition(game, ["p2"]);
-      expect(result?.winner).toBe(WerewolfWinner.Spoiler);
-    });
-
-    it("dead Spoiler does not affect outcome", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-      ]);
-      // Spoiler is dead; Werewolf dead → Village wins normally
-      const result = checkWinCondition(game, ["p1", "p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-  });
-
-  describe("Draw", () => {
-    it("draw when all players are simultaneously eliminated", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1", "p2"]);
-      expect(result?.type).toBe(GameStatus.Finished);
-      expect(result?.winner).toBe(WerewolfWinner.Draw);
-    });
-
-    it("draw when wolf and chupacabra both die on the same night", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // p3 already dead; p1 and p2 kill each other this night
-      const result = checkWinCondition(game, ["p3", "p1", "p2"]);
-      expect(result?.winner).toBe(WerewolfWinner.Draw);
-    });
-
-    it("alive Spoiler overrides Draw result", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-      ]);
-      // Wolf and Villager both dead — would be Draw, but Spoiler is alive
-      const result = checkWinCondition(game, ["p1", "p2"]);
-      expect(result?.winner).toBe(WerewolfWinner.Spoiler);
-    });
-  });
-
-  describe("Neutral roles in wolf-vs-village balance", () => {
-    it("Tanner alive prevents wolves from winning by being uncounted", () => {
-      // 1 wolf + 1 villager + 1 tanner: without counting Tanner, wolves would win (1v1)
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Tanner },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("Executioner alive prevents wolves from winning by being uncounted", () => {
-      // 1 wolf + 1 villager + 1 executioner: without counting Executioner, wolves would win (1v1)
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Executioner },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("Dracula alive prevents wolves from winning by being uncounted", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Dracula },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("Zombie alive prevents wolves from winning by being uncounted", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Zombie },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-  });
-
-  describe("Zombie", () => {
-    it("Zombie wins when infected alive outnumber healthy alive", () => {
-      // zombie + p1 (infected) + p2 (infected) + p3 (healthy): 2 infected > 1 healthy
-      const ts = makeDayTurnState({
-        roleState: { zombie: { infected: ["p1", "p2"] } },
-      });
-      const game = makeGame(
-        [
-          { playerId: "zombie", roleDefinitionId: WerewolfRole.Zombie },
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        ts,
-      );
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Zombie);
-    });
-
-    it("game continues when infected equals healthy", () => {
-      const ts = makeDayTurnState({
-        roleState: { zombie: { infected: ["p1"] } },
-      });
-      const game = makeGame(
-        [
-          { playerId: "zombie", roleDefinitionId: WerewolfRole.Zombie },
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        ts,
-      );
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).not.toBe(WerewolfWinner.Zombie);
-    });
-
-    it("Zombie does not win when dead", () => {
-      const ts = makeDayTurnState({
-        roleState: { zombie: { infected: ["p1", "p2"] } },
-      });
-      const game = makeGame(
-        [
-          { playerId: "zombie", roleDefinitionId: WerewolfRole.Zombie },
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        ts,
-      );
-      const result = checkWinCondition(game, ["zombie"]);
-      expect(result?.winner).not.toBe(WerewolfWinner.Zombie);
-    });
-
-    it("dead infected players are not counted for Zombie win", () => {
-      // zombie + p1 (infected, dead) + p2 (infected) + p3 (healthy)
-      // alive: 1 infected vs 1 healthy → no win
-      const ts = makeDayTurnState({
-        roleState: { zombie: { infected: ["p1", "p2"] } },
-      });
-      const game = makeGame(
-        [
-          { playerId: "zombie", roleDefinitionId: WerewolfRole.Zombie },
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        ts,
-      );
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).not.toBe(WerewolfWinner.Zombie);
-    });
-  });
-
-  describe("Illuminati", () => {
-    it("Illuminati wins when alive and ≤3 players remain during a wolf win", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Illuminati },
-      ]);
-      // 1 wolf + 1 illuminati: wolves would win (1 bad >= 1 non-bad), but Illuminati is alive with ≤3
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Illuminati);
-    });
-
-    it("Illuminati wins when alive and ≤3 players remain during a village win", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Illuminati },
-      ]);
-      // Wolf dead → Village wins, but Illuminati is alive with ≤3 remaining
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Illuminati);
-    });
-
-    it("Illuminati does not win when more than 3 players are alive", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Illuminati },
-      ]);
-      // 2 wolves + 1 villager + 1 illuminati = 4 alive: wolves win (2 bad >= 2 non-bad)
-      // but 4 > 3, so Illuminati does NOT override
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Werewolves);
-    });
-
-    it("dead Illuminati does not affect outcome", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Illuminati },
-      ]);
-      // Illuminati is dead; Wolf dead → Village wins normally
-      const result = checkWinCondition(game, ["p1", "p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-
-    it("Illuminati counts as neutral and prevents premature wolf win", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Illuminati },
-      ]);
-      // 1 wolf vs 1 villager + 1 illuminati: wolves do NOT win yet (1 < 2)
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("Illuminati wins over Spoiler when both are alive and ≤3 remain", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Illuminati },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-      ]);
-      // Wolf wins (1 bad >= 1 non-bad with only Illuminati as neutral), 3 players alive
-      // Illuminati override fires before Spoiler
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Illuminati);
-    });
-  });
-
-  describe("Arsonist", () => {
-    it("Arsonist wins when all Bad are dead and ≤1 Good remains", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result?.winner).toBe(WerewolfWinner.Arsonist);
-    });
-
-    it("Arsonist wins when all Bad are dead and 0 Good remain", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1", "p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.Arsonist);
-    });
-
-    it("game continues when Arsonist is alive and >1 Good remain", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result).toBeUndefined();
-    });
-
-    it("game continues when Bad and Arsonist are the only survivors", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // Only wolf and arsonist remain — conflicting win conditions
-      const result = checkWinCondition(game, ["p3"]);
-      expect(result).toBeUndefined();
-    });
-
-    it("game continues when both Arsonist and Chupacabra are alive with ≤1 Good", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Chupacabra },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // Wolf dead, both neutral killers alive with 1 Good — neither can win yet
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result).toBeUndefined();
-    });
-
-    it("Arsonist counts as a non-Bad threat for Werewolf win condition", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // 1 wolf vs 1 arsonist + 1 villager: wolves do NOT win (1 < 2)
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("Village does not win while Arsonist is alive", () => {
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Arsonist },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-      ]);
-      // Wolf dead, Arsonist + 2 Good remain — Arsonist still has >1 Good to overcome, game continues
-      const result = checkWinCondition(game, ["p1"]);
-      expect(result).toBeUndefined();
-    });
-  });
-
-  describe("Mercenary win condition", () => {
-    it("Mercenary alive prevents wolves from winning by being uncounted", () => {
-      // 1 wolf + 1 villager + 1 mercenary: Mercenary counts as neutral, wolves don't win yet
-      const game = makeGame([
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Mercenary },
-      ]);
-      const result = checkWinCondition(game, []);
-      expect(result).toBeUndefined();
-    });
-
-    it("Mercenary co-wins when Village wins and a bribed Good player is alive", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // Wolf dead → Village wins; Mercenary alive with bribed Good player alive → co-win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-
-    it("Mercenary co-wins when Werewolves win and a bribed Bad player is alive", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p1"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // 1 wolf, 1 neutral Mercenary — wolves outnumber non-bad → wolves win; Mercenary co-wins
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Werewolves);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-
-    it("Mercenary does not win when bribed player is dead", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // Wolf dead → Village wins; but bribed player p3 is dead → Mercenary doesn't win
-      const result = checkWinCondition(game, ["p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-      expect(result?.victoryConditionKey).toBeUndefined();
-    });
-
-    it("Mercenary does not win when they have no bribed players", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-        ],
-        makeDayTurnState({ deadPlayerIds: [] }),
-      );
-      // Wolf dead → Village wins; but Mercenary has no bribed players → no win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-
-    it("Mercenary does not win when they are dead", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: ["p2"],
-        }),
-      );
-      // Wolf dead → Village wins; Mercenary dead → Mercenary doesn't win
-      const result = checkWinCondition(game, ["p2"]);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-
-    it("Mercenary does not win when bribed player is on wrong team", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Tanner },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // Wolf dead → Village wins (p3 Tanner is neutral, not Team.Good) → Mercenary doesn't win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Village);
-    });
-
-    it("Spoiler wins alone when Spoiler is alive and bribed player is not the Spoiler", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-          { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p4"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // Village wins; Spoiler alive → Spoiler wins; Mercenary bribed p4 (Villager, not Spoiler) → no co-win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Spoiler);
-      expect(result?.victoryConditionKey).toBeUndefined();
-    });
-
-    it("Mercenary co-wins with Spoiler when Spoiler is bribed and alive", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Spoiler },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // Village wins; Spoiler alive → Spoiler wins; Mercenary bribed the Spoiler → co-win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Spoiler);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-
-    it("Mercenary co-wins with Illuminati when Illuminati is bribed and alive", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Illuminati },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // ≤3 players, win fires → Illuminati wins; Mercenary bribed the Illuminati → co-win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Illuminati);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-
-    it("Mercenary co-wins with Lone Wolf when Lone Wolf is bribed and alive", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.LoneWolf },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p1"] } },
-          deadPlayerIds: [],
-        }),
-      );
-      // Lone Wolf wins; Mercenary bribed the Lone Wolf → co-win
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.LoneWolf);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-
-    it("Mercenary does not co-win on Lone Wolf victory when bribed player is dead", () => {
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.LoneWolf },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p3"] } },
-          deadPlayerIds: ["p3"],
-        }),
-      );
-      // Lone Wolf wins (only Bad is LoneWolf, matches Mercenary neutral); bribed p3 is dead → no co-win
-      const result = checkWinCondition(game, ["p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.LoneWolf);
-      expect(result?.victoryConditionKey).toBeUndefined();
-    });
-
-    it("Mercenary co-wins with Zombie when the Zombie is bribed and wins", () => {
-      // p1=Zombie, p2=Villager(infected), p3=Villager(infected), p4=Mercenary
-      // infected(2) > healthy(0) → Zombie wins; Mercenary bribed Zombie → co-win
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Zombie },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-          { playerId: "p4", roleDefinitionId: WerewolfRole.Mercenary },
-        ],
-        makeDayTurnState({
-          roleState: {
-            zombie: { infected: ["p2", "p3"] },
-            mercenary: { charged: false, bribedPlayerIds: ["p1"] },
-          },
-          deadPlayerIds: [],
-        }),
-      );
-      const result = checkWinCondition(game, []);
-      expect(result?.winner).toBe(WerewolfWinner.Zombie);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-
-    it("Mercenary co-wins with Arsonist when the Arsonist is bribed and wins", () => {
-      // p1=Arsonist, p2=Mercenary; Bad dead → Arsonist wins (≤1 Good remains);
-      // Mercenary bribed the Arsonist → co-win
-      const game = makeGame(
-        [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Arsonist },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Mercenary },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-        ],
-        makeDayTurnState({
-          roleState: { mercenary: { charged: false, bribedPlayerIds: ["p1"] } },
-          deadPlayerIds: ["p3"],
-        }),
-      );
-      // Arsonist + Mercenary remain; ≤1 Good (p3 dead) → Arsonist wins; Mercenary bribed Arsonist → co-win
-      const result = checkWinCondition(game, ["p3"]);
-      expect(result?.winner).toBe(WerewolfWinner.Arsonist);
-      expect(result?.victoryConditionKey).toBe(WerewolfWinner.Mercenary);
-    });
-  });
-});
